+++ src/main/java/songstress/cards/AngelicLuck.java
@@ -1,0 +1,51 @@
+package songstress.cards;
+
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import songstress.actions.AngelicLuckAction;
+
+public class AngelicLuck extends AbstractSongstressCard {
+
+	public static final String ID = "AngelicLuck";
+	private static final CardStrings cardStrings = getCardStrings(ID);
+	public static final String NAME = cardStrings.NAME;
+	private static final int COST = 3;
+	private static final int COST_UPGRADED = 2;
+	public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+	private static final int CARDS = 3;
+	private static final int UPGRADE_CARDS = 2;
+	private static final CardType TYPE = CardType.SKILL;
+	private static final CardRarity RARITY = CardRarity.RARE;
+	private static final CardTarget TARGET = CardTarget.SELF;
+
+	public AngelicLuck() {
+		super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, TARGET);
+		magicNumber = baseMagicNumber = CARDS;
+		exhaust = true;
+	}
+
+	@Override
+	public void upgrade() {
+		if (!upgraded) {
+			upgradeName();
+			upgradeMagicNumber(UPGRADE_CARDS);
+			upgradeBaseCost(COST_UPGRADED);
+		}
+	}
+
+	@Override
+	public AbstractCard makeCopy() {
+		return new AngelicLuck();
+	}
+
+	@Override
+	public void use(AbstractPlayer player, AbstractMonster monster) {
+		AbstractDungeon.actionManager
+				.addToBottom(new AngelicLuckAction(AbstractDungeon.getRandomMonster(), magicNumber));
+	}
+
+}