+++ src/main/java/songstress/cards/Purge.java
@@ -1,0 +1,48 @@
+package songstress.cards;
+
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import songstress.actions.PurgeAction;
+
+public class Purge extends AbstractSongstressCard {
+
+	public static final String ID = "Purge";
+	private static final CardStrings cardStrings = getCardStrings(ID);
+	public static final String NAME = cardStrings.NAME;
+	private static final int COST = 2;
+	public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+	private static final int STRENGTH_GAIN = 1;
+	private static final int UPGRADE_STRENGTH_GAIN = 1;
+	private static final CardType TYPE = CardType.SKILL;
+	private static final CardRarity RARITY = CardRarity.RARE;
+	private static final CardTarget TARGET = CardTarget.SELF;
+
+	public Purge() {
+		super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, TARGET);
+		baseMagicNumber = magicNumber = STRENGTH_GAIN;
+		exhaust = true;
+	}
+
+	@Override
+	public void upgrade() {
+		if (!upgraded) {
+			upgradeName();
+			upgradeMagicNumber(UPGRADE_STRENGTH_GAIN);
+		}
+	}
+
+	@Override
+	public AbstractCard makeCopy() {
+		return new Purge();
+	}
+
+	@Override
+	public void use(AbstractPlayer p, AbstractMonster m) {
+		AbstractDungeon.actionManager.addToBottom(new PurgeAction(magicNumber));
+	}
+
+}