+++ src/main/java/songstress/cards/Revival.java
@@ -1,0 +1,47 @@
+package songstress.cards;
+
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import songstress.actions.RevivalAction;
+
+public class Revival extends AbstractSongstressCard {
+
+	public static final String ID = "Revival";
+	private static final CardStrings cardStrings = getCardStrings(ID);
+	public static final String NAME = cardStrings.NAME;
+	private static final int COST = 2;
+	public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+	private static final int CARDS = 2;
+	private static final int UPGRADE_CARDS = 1;
+	private static final CardType TYPE = CardType.SKILL;
+	private static final CardRarity RARITY = CardRarity.RARE;
+	private static final CardTarget TARGET = CardTarget.SELF;
+
+	public Revival() {
+		super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, TARGET);
+		magicNumber = baseMagicNumber = CARDS;
+	}
+
+	@Override
+	public void upgrade() {
+		if (!upgraded) {
+			upgradeName();
+			upgradeMagicNumber(UPGRADE_CARDS);
+		}
+	}
+
+	@Override
+	public AbstractCard makeCopy() {
+		return new Revival();
+	}
+
+	@Override
+	public void use(AbstractPlayer player, AbstractMonster monster) {
+		AbstractDungeon.actionManager.addToBottom(new RevivalAction(false, magicNumber));
+	}
+
+}